Added How to Use paragraph
Added How to Use paragraph located in strings.xml. This string can be
edited as we see fit. For now I just placed it in the bottom of the
send_fragment as it's just empty space right now. Ideally we would want
this to be located in the Options Menu, where I've also created a "How
to Use" selection tab.