Added FileViewer class, changed Receive
Added FileViewer class that contains openFile() function. This function
opens a file by creating an Intent of the appropriate type given the
file extension of the file. This should be a straightforward, easy way
to open files using Android's default applications. Usage of this class
is pretty simple and can easily be added for viewing documents in the
DocumentListFragment.

Also added mFileDest variable to hold the path to the file after it is
saved locally to the apps external storage on the SD card.

Javadocs already included, I think they're the correct format.